HACK: scripts: Makefile: disable size_append
Drivers in techpack/ are supposed to be built externally as modules.
When they are built in-tree. They trigger file-size.sh for all kernel
headers. As a result, immense amount of shells are spawned by size_append.

It is time-consuming and resource-hungry to spawn processes and repeatly
call "ls" to get size of file. Plus, result is not cached for incremental
builds. This significantly increased the incremental compilation time.

As we are not using size_append for this kernel because kernel image
compression is not enabled, let's just disable size_append.

Change-Id: Ia7d80d520a1e7d4c54c08658ae24f1ee5b772e36
Signed-off-by: Jesse Chan <jc@linux.com>